requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.7
flower==2.0.1
gunicorn==23.0.0
humanize==4.11.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
packaging==24.1
pillow==10.4.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.38.4
